+++ CopilotKit/packages/react-core/src/hooks/use-chat.ts
@@ -16,6 +16,7 @@ import {
   MessageStatusCode,
   MessageRole,
   Role,
+  CopilotRequestType,
 } from "@copilotkit/runtime-client-gql";
 
 import { CopilotApiConfig } from "../context";
@@ -133,8 +134,8 @@ export function useChat(options: UseChatOptions): UseChatHelpers {
     const messagesWithContext = [systemMessage, ...(initialMessages || []), ...previousMessages];
 
     const stream = CopilotRuntimeClient.asStream(
-      runtimeClient.generateCopilotResponse(
-        {
+      runtimeClient.generateCopilotResponse({
+        data: {
           frontend: {
             actions: actions.map((action) => ({
               name: action.name,
@@ -148,19 +149,28 @@ export function useChat(options: UseChatOptions): UseChatHelpers {
           ...(copilotConfig.cloud
             ? {
                 cloud: {
-                  guardrails: {
-                    inputValidationRules: {
-                      allowList: copilotConfig.cloud.guardrails.input.restrictToTopic.validTopics,
-                      denyList: copilotConfig.cloud.guardrails.input.restrictToTopic.invalidTopics,
-                    },
-                  },
+                  ...(copilotConfig.cloud.guardrails?.input?.restrictToTopic?.enabled
+                    ? {
+                        guardrails: {
+                          inputValidationRules: {
+                            allowList:
+                              copilotConfig.cloud.guardrails.input.restrictToTopic.validTopics,
+                            denyList:
+                              copilotConfig.cloud.guardrails.input.restrictToTopic.invalidTopics,
+                          },
+                        },
+                      }
+                    : {}),
                 },
               }
             : {}),
+          metadata: {
+            requestType: CopilotRequestType.Chat,
+          },
         },
-        copilotConfig.properties,
-        abortControllerRef.current?.signal,
-      ),
+        properties: copilotConfig.properties,
+        signal: abortControllerRef.current?.signal,
+      }),
     );
 
     const guardrailsEnabled =