+++ CopilotKit/packages/runtime/src/graphql/resolvers/copilot.resolver.ts
@@ -16,7 +16,6 @@ import { CopilotResponse } from "../types/copilot-response.type";
 import { MessageRole } from "../types/enums";
 import { Repeater } from "graphql-yoga";
 import type { CopilotRequestContextProperties, GraphQLContext } from "../../lib/integrations";
-import { nanoid } from "nanoid";
 import { RuntimeEvent, RuntimeEventTypes } from "../../service-adapters/events";
 import {
   FailedMessageStatus,
@@ -26,7 +25,7 @@ import {
 import { ResponseStatusUnion, SuccessResponseStatus } from "../types/response-status.type";
 import { GraphQLJSONObject } from "graphql-scalars";
 import { plainToInstance } from "class-transformer";
-import { GuardrailsResult, GuardrailsResultStatus } from "../types/guardrails-result.type";
+import { GuardrailsResult } from "../types/guardrails-result.type";
 import { GraphQLError } from "graphql";
 import {
   GuardrailsValidationFailureResponse,
@@ -34,6 +33,8 @@ import {
   UnknownErrorResponse,
 } from "../../utils";
 import { ActionExecutionMessage, Message, ResultMessage, TextMessage } from "../types/converted";
+import telemetry from "../../lib/telemetry-client";
+import { randomId } from "@copilotkit/shared";
 
 const invokeGuardrails = async ({
   baseUrl,
@@ -48,6 +49,8 @@ const invokeGuardrails = async ({
   onResult: (result: GuardrailsResult) => void;
   onError: (err: Error) => void;
 }) => {
+  console.log("invokeGuardrails.baseUrl", baseUrl);
+
   if (
     data.messages.length &&
     data.messages[data.messages.length - 1].textMessage?.role === MessageRole.user
@@ -105,6 +108,11 @@ export class CopilotResolver {
     @Arg("properties", () => GraphQLJSONObject, { nullable: true })
     properties?: CopilotRequestContextProperties,
   ) {
+    telemetry.capture("oss.runtime.copilot_request_created", {
+      "cloud.guardrails.enabled": data.cloud?.guardrails !== undefined,
+      requestType: data.metadata.requestType,
+    });
+
     let logger = ctx.logger.child({ component: "CopilotResolver.generateCopilotResponse" });
     logger.debug({ data }, "Generating Copilot response");
 
@@ -133,14 +141,15 @@ export class CopilotResolver {
 
       if (process.env.COPILOT_CLOUD_BASE_URL) {
         copilotCloudBaseUrl = process.env.COPILOT_CLOUD_BASE_URL;
-      } else if (ctx._copilotkit.baseUrl) {
-        copilotCloudBaseUrl = ctx._copilotkit.baseUrl;
+      } else if (ctx._copilotkit.cloud?.baseUrl) {
+        copilotCloudBaseUrl = ctx._copilotkit.cloud?.baseUrl;
       } else {
         copilotCloudBaseUrl = "https://api.cloud.copilotkit.ai";
       }
 
       logger = logger.child({ copilotCloudBaseUrl });
     }
+
     logger.debug("Setting up subjects");
     const responseStatus$ = new ReplaySubject<typeof ResponseStatusUnion>();
     const interruptStreaming$ = new ReplaySubject<{ reason: string; messageId?: string }>();
@@ -158,7 +167,7 @@ export class CopilotResolver {
     logger.debug("Processing");
     const {
       eventSource,
-      threadId = nanoid(),
+      threadId = randomId(),
       runId,
       actions,
     } = await copilotRuntime.process({
@@ -206,7 +215,7 @@ export class CopilotResolver {
                 // resolve messages promise to the middleware
                 outputMessages = [
                   plainToInstance(TextMessage, {
-                    id: nanoid(),
+                    id: randomId(),
                     createdAt: new Date(),
                     content: result.reason,
                     role: MessageRole.assistant,
@@ -269,7 +278,7 @@ export class CopilotResolver {
                 // signal when we are done streaming
                 const streamingTextStatus = new Subject<typeof MessageStatusUnion>();
 
-                const messageId = nanoid();
+                const messageId = randomId();
 
                 // push the new message
                 pushMessage({
@@ -404,7 +413,7 @@ export class CopilotResolver {
               case RuntimeEventTypes.ActionExecutionResult:
                 logger.debug({ result: event.result }, "Action execution result event received");
                 pushMessage({
-                  id: nanoid(),
+                  id: randomId(),
                   status: new SuccessMessageStatus(),
                   createdAt: new Date(),
                   actionExecutionId: event.actionExecutionId,
@@ -414,7 +423,7 @@ export class CopilotResolver {
 
                 outputMessages.push(
                   plainToInstance(ResultMessage, {
-                    id: nanoid(),
+                    id: randomId(),
                     createdAt: new Date(),
                     actionExecutionId: event.actionExecutionId,
                     actionName: event.actionName,