+++ CopilotKit/packages/runtime/src/lib/telemetry-client.ts
@@ -1,0 +1,9 @@
+import { TelemetryClient } from "@copilotkit/shared";
+const packageJson = require("../../package.json");
+
+const telemetryClient = new TelemetryClient({
+  packageName: packageJson.name,
+  packageVersion: packageJson.version,
+});
+
+export default telemetryClient;