+++ CopilotKit/packages/shared/src/utils/index.ts
@@ -1,2 +1,2 @@
-export * from "./annotated-function";
 export * from "./json-schema";
+export * from "./random-id";