+++ CopilotKit/examples/node-express/CHANGELOG.md
@@ -1,5 +1,83 @@
 # node
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/runtime@1.0.5
+  - @copilotkit/shared@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- Remove nanoid
+- Updated dependencies
+  - @copilotkit/runtime@1.0.4
+  - @copilotkit/shared@1.0.4
+
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/runtime@1.0.3
+  - @copilotkit/shared@1.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/runtime@1.0.2
+  - @copilotkit/shared@1.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/runtime@1.0.1
+  - @copilotkit/shared@1.0.1
+
+## 1.0.0
+
+### Major Changes
+
+- b6a4b6eb: V1.0 Release Candidate
+
+  - A robust new protocol between the frontend and the Copilot Runtime
+  - Support for Copilot Cloud
+  - Generative UI
+  - Support for LangChain universal tool calling
+  - OpenAI assistant API streaming
+
+- V1.0 Release
+
+  - A robust new protocol between the frontend and the Copilot Runtime
+  - Support for Copilot Cloud
+  - Generative UI
+  - Support for LangChain universal tool calling
+  - OpenAI assistant API streaming
+
+### Patch Changes
+
+- Updated dependencies [b6a4b6eb]
+- Updated dependencies [b6a4b6eb]
+- Updated dependencies [b6a4b6eb]
+- Updated dependencies
+  - @copilotkit/runtime@1.0.0
+  - @copilotkit/shared@1.0.0
+
+## 1.0.0-beta.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/runtime@1.0.0-beta.2
+  - @copilotkit/shared@1.0.0-beta.2
+
 ## 1.23.0
 
 ### Minor Changes