+++ CopilotKit/packages/runtime/CHANGELOG.md
@@ -1,1 +1,104 @@
 # @copilotkit/runtime
+
+## 1.0.5
+
+### Patch Changes
+
+- Include @copilotkit/runtime-client-gql NPM package version in request to Runtime
+- Updated dependencies
+  - @copilotkit/shared@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- Remove nanoid
+- Updated dependencies
+  - @copilotkit/shared@1.0.4
+
+## 1.0.3
+
+### Patch Changes
+
+- Add README.md to published packages and add keywords to package.json
+- Updated dependencies
+  - @copilotkit/shared@1.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- Add README.md and homepage/url to published packages
+- Updated dependencies
+  - @copilotkit/shared@1.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- Remove PostHog, use Segment Anonymous Telemetry instead
+- Updated dependencies
+  - @copilotkit/shared@1.0.1
+
+## 1.0.0
+
+### Major Changes
+
+- b6a4b6eb: V1.0 Release Candidate
+
+  - A robust new protocol between the frontend and the Copilot Runtime
+  - Support for Copilot Cloud
+  - Generative UI
+  - Support for LangChain universal tool calling
+  - OpenAI assistant API streaming
+
+- V1.0 Release
+
+  - A robust new protocol between the frontend and the Copilot Runtime
+  - Support for Copilot Cloud
+  - Generative UI
+  - Support for LangChain universal tool calling
+  - OpenAI assistant API streaming
+
+### Patch Changes
+
+- b6a4b6eb: Introduce anonymous telemetry
+- b6a4b6eb: Set default Copilot Cloud runtime URL to versioned URL (v1)
+- Updated dependencies [b6a4b6eb]
+- Updated dependencies [b6a4b6eb]
+- Updated dependencies [b6a4b6eb]
+- Updated dependencies
+  - @copilotkit/shared@1.0.0
+
+## 1.0.0-beta.2
+
+### Patch Changes
+
+- Set default Copilot Cloud runtime URL to versioned URL (v1)
+- Updated dependencies
+  - @copilotkit/shared@1.0.0-beta.2
+
+## 1.0.0-beta.1
+
+### Patch Changes
+
+- Introduce anonymous telemetry
+- Updated dependencies
+  - @copilotkit/shared@1.0.0-beta.1
+
+## 1.0.0-beta.0
+
+### Major Changes
+
+- V1.0 Release Candidate
+
+  - A robust new protocol between the frontend and the Copilot Runtime
+  - Support for Copilot Cloud
+  - Generative UI
+  - Support for LangChain universal tool calling
+  - OpenAI assistant API streaming
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/shared@1.0.0-beta.0